Several changes (see description)
Enemies now have a EnemyType enum. This type is assigned in the Init method of each enemy-type class.
Added a virtual empty method in Enemy class for actions that must be done when the player and the enemy collide. This method is implemented in each enemy-type class (depending they need it or not) by overriding the parents class' method.
Finished Frog enemy-type implementation.
Started Stomp enemy-type implementation.
Updated the map a little.
Other minor changes.