Improved how/what PickUps apply  when picking them up (see description) and other minor changes
Before, the PickUps had a method that received the Player as parameter and it was them who changed the Player's info.
Now, the PickUps returns specific info so the World knows how to affect the player (sum: Now is the World who modifies the player when picking up a Gem or PowerUp).
Created two methods in the Player that are called when picking up a pickable: addScore, addLife and applySpeedBoost.
Modified the effect when picking up a ExtraLife: Now the maximum lives amount is 2, and picking up a life while having 2 lives adds 10 points to the player's score.
Modified the effect when picking up a SpeedBoost: Now the speed boost is temporal and stays for 10 seconds.